Rename api module to client
