+++ pages/api/vehicles/[vehicleQuery].js
@@ -1,4 +1,4 @@
-import { fetchPrice } from "../../../api/fipe"
+import { fetchPrice } from "../../../client/fipe"
 
 export default async function handler(req, res) {
   if (req.method !== 'GET') {
